+++ graal-js/src/com.oracle.js.parser/src/com/oracle/js/parser/Parser.java
@@ -67,6 +67,7 @@ import static com.oracle.js.parser.TokenType.RBRACKET;
 import static com.oracle.js.parser.TokenType.RPAREN;
 import static com.oracle.js.parser.TokenType.SEMICOLON;
 import static com.oracle.js.parser.TokenType.SET;
+import static com.oracle.js.parser.TokenType.SPREAD_ARGUMENT;
 import static com.oracle.js.parser.TokenType.SPREAD_ARRAY;
 import static com.oracle.js.parser.TokenType.SPREAD_OBJECT;
 import static com.oracle.js.parser.TokenType.STATIC;
@@ -181,6 +182,7 @@ public class Parser extends AbstractParser {
     private static final String ARROW_FUNCTION_NAME = ":=>";
 
     private static final String FUNCTION_PARAMETER_CONTEXT = "function parameter";
+    private static final String CATCH_PARAMETER_CONTEXT = "catch parameter";
 
     private static final boolean ES6_FOR_OF = Options.getBooleanProperty("parser.for.of", true);
     private static final boolean ES6_CLASS = Options.getBooleanProperty("parser.class", true);
@@ -198,6 +200,7 @@ public class Parser extends AbstractParser {
     private static final boolean ES8_ASYNC_FUNCTION = Options.getBooleanProperty("parser.async.function", true);
     private static final boolean ES8_REST_SPREAD_PROPERTY = Options.getBooleanProperty("parser.rest.spread.property", true);
     private static final boolean ES8_FOR_AWAIT_OF = Options.getBooleanProperty("parser.for.await.of", true);
+    private static final boolean ES2019_OPTIONAL_CATCH_BINDING = Options.getBooleanProperty("parser.optional.catch.binding", true);
 
     private static final int REPARSE_IS_PROPERTY_ACCESSOR = 1 << 0;
     private static final int REPARSE_IS_METHOD = 1 << 1;
@@ -2886,17 +2889,22 @@ loop:
         // Create try.
 
         try {
-            final Block       tryBody     = getBlock(true);
+            final Block tryBody = getBlock(true);
             final ArrayList<Block> catchBlocks = new ArrayList<>();
 
             while (type == CATCH) {
                 final int  catchLine  = line;
                 final long catchToken = token;
                 next();
+
+                if (type == LBRACE && ES2019_OPTIONAL_CATCH_BINDING) {
+                    catchBlocks.add(catchBlock(catchToken, catchLine, null, null, null));
+                    break;
+                }
+
                 expect(LPAREN);
 
-                final String contextString = "catch parameter";
-                final Expression catchParameter = bindingIdentifierOrPattern(contextString);
+                final Expression catchParameter = bindingIdentifierOrPattern(CATCH_PARAMETER_CONTEXT);
                 final IdentNode exception;
                 final Expression pattern;
                 if (catchParameter instanceof IdentNode) {
@@ -2921,29 +2929,7 @@ loop:
 
                 expect(RPAREN);
 
-                final ParserContextBlockNode catchBlock = newBlock();
-                try {
-                    appendStatement(new VarNode(catchLine, Token.recast(exception.getToken(), LET), exception.getFinish(), exception.setIsDeclaredHere(), null, VarNode.IS_LET));
-                    if (pattern != null) {
-                        verifyDestructuringBindingPattern(pattern, new Consumer<IdentNode>() {
-                            @Override
-                            public void accept(IdentNode identNode) {
-                                verifyStrictIdent(identNode, contextString);
-                                final int varFlags = VarNode.IS_LET | VarNode.IS_DESTRUCTURING;
-                                final VarNode var = new VarNode(catchLine, Token.recast(identNode.getToken(), LET), identNode.getFinish(), identNode.setIsDeclaredHere(), null, varFlags);
-                                appendStatement(var);
-                            }
-                        });
-                    }
-
-                    // Get CATCH body.
-                    final Block catchBody = getBlock(true);
-                    final CatchNode catchNode = new CatchNode(catchLine, catchToken, finish, exception, pattern, ifExpression, catchBody, false);
-                    appendStatement(catchNode);
-                } finally {
-                    restoreBlock(catchBlock);
-                    catchBlocks.add(new Block(catchBlock.getToken(), Math.max(finish, Token.descPosition(catchBlock.getToken())), catchBlock.getFlags() | Block.IS_SYNTHETIC, catchBlock.getStatements()));
-                }
+                catchBlocks.add(catchBlock(catchToken, catchLine, exception, pattern, ifExpression));
 
                 // If unconditional catch then should to be the end.
                 if (ifExpression == null) {
@@ -2973,6 +2959,34 @@ loop:
         }
 
         appendStatement(new BlockStatement(startLine, new Block(tryToken, finish, outer.getFlags() | Block.IS_SYNTHETIC, outer.getStatements())));
+    }
+
+    private Block catchBlock(final long catchToken, final int catchLine, final IdentNode exception, final Expression pattern, final Expression ifExpression) {
+        final ParserContextBlockNode catchBlock = newBlock();
+        try {
+            if (exception != null) {
+                appendStatement(new VarNode(catchLine, Token.recast(exception.getToken(), LET), exception.getFinish(), exception.setIsDeclaredHere(), null, VarNode.IS_LET));
+                if (pattern != null) {
+                    verifyDestructuringBindingPattern(pattern, new Consumer<IdentNode>() {
+                        @Override
+                        public void accept(IdentNode identNode) {
+                            verifyStrictIdent(identNode, CATCH_PARAMETER_CONTEXT);
+                            final int varFlags = VarNode.IS_LET | VarNode.IS_DESTRUCTURING;
+                            final VarNode var = new VarNode(catchLine, Token.recast(identNode.getToken(), LET), identNode.getFinish(), identNode.setIsDeclaredHere(), null, varFlags);
+                            appendStatement(var);
+                        }
+                    });
+                }
+            }
+
+            // Get CATCH body.
+            final Block catchBody = getBlock(true);
+            final CatchNode catchNode = new CatchNode(catchLine, catchToken, finish, exception, pattern, ifExpression, catchBody, false);
+            appendStatement(catchNode);
+        } finally {
+            restoreBlock(catchBlock);
+        }
+        return new Block(catchBlock.getToken(), Math.max(finish, Token.descPosition(catchBlock.getToken())), catchBlock.getFlags() | Block.IS_SYNTHETIC, catchBlock.getStatements());
     }
 
     /**
@@ -5201,7 +5215,7 @@ loop:
         if (paramListExpr == null) {
             // empty parameter list, i.e. () =>
             return;
-        } else if (paramListExpr instanceof IdentNode || paramListExpr.isTokenType(ASSIGN) || isDestructuringLhs(paramListExpr)) {
+        } else if (paramListExpr instanceof IdentNode || paramListExpr.isTokenType(ASSIGN) || isDestructuringLhs(paramListExpr) || paramListExpr.isTokenType(SPREAD_ARGUMENT)) {
             convertArrowParameter(paramListExpr, 0, functionLine, function);
         } else if (paramListExpr instanceof BinaryNode && Token.descType(paramListExpr.getToken()) == COMMARIGHT) {
             ArrayList<Expression> params = new ArrayList<>();
@@ -5214,7 +5228,12 @@ loop:
             params.add(car);
 
             for (int i = params.size() - 1, pos = 0; i >= 0; i--, pos++) {
-                convertArrowParameter(params.get(i), pos, functionLine, function);
+                Expression param = params.get(i);
+                if (i != 0 && param.isTokenType(SPREAD_ARGUMENT)) {
+                    throw error(AbstractParser.message("invalid.arrow.parameter"), param.getToken());
+                } else {
+                    convertArrowParameter(param, pos, functionLine, function);
+                }
             }
         } else {
             throw error(AbstractParser.message("expected.arrow.parameter"), paramListExpr.getToken());
@@ -5269,9 +5288,47 @@ loop:
             if (currentFunction != null) {
                 addDestructuringParameter(paramToken, param.getFinish(), paramLine, param, null, currentFunction);
             }
+        } else if (param.isTokenType(SPREAD_ARGUMENT)) {
+            Expression expression = ((UnaryNode) param).getExpression();
+            if (expression instanceof IdentNode && identAtTheEndOfArrowParamList()) {
+                IdentNode rest = ((IdentNode) expression).setIsRestParameter();
+                convertArrowParameter(rest, index, paramLine, currentFunction);
+            } else {
+                throw error(AbstractParser.message("invalid.arrow.parameter"), param.getToken());
+            }
         } else {
             throw error(AbstractParser.message("invalid.arrow.parameter"), param.getToken());
         }
+    }
+
+    // Checks wheter there is IDENT and RPAREN before ARROW. The function
+    // is used to verify that rest parameter is not followed by a trailing comma.
+    private boolean identAtTheEndOfArrowParamList() {
+        int idx = k - 1;
+        assert T(idx) == ARROW;
+        while (true) {
+            idx--;
+            TokenType t = T(idx);
+            if (t == COMMENT) {
+                continue;
+            } else if (t == RPAREN) {
+                break;
+            } else {
+                return false;
+            }
+        }
+        while (true) {
+            idx--;
+            TokenType t = T(idx);
+            if (t == COMMENT) {
+                continue;
+            } else if (t == IDENT) {
+                break;
+            } else {
+                return false;
+            }
+        }
+        return true;
     }
 
     private boolean checkNoLineTerminator() {