+++ graal-js/src/com.oracle.truffle.js.scriptengine/src/com/oracle/truffle/js/scriptengine/GraalJSScriptEngine.java
@@ -92,7 +92,7 @@ public final class GraalJSScriptEngine extends AbstractScriptEngine implements C
         Context.Builder contextConfigToUse = contextConfig;
         if (contextConfigToUse == null) {
             // default config
-            contextConfigToUse = Context.newBuilder(ID).allowHostAccess(true).allowCreateThread(true);
+            contextConfigToUse = Context.newBuilder(ID).allowHostAccess(true).allowCreateThread(true).option("js.nashorn-compat", "true").option("js.syntax-extensions", "true");
         }
         this.factory = new GraalJSEngineFactory(engineToUse);
         this.contextConfig = contextConfigToUse.engine(engineToUse);
@@ -326,7 +326,9 @@ public final class GraalJSScriptEngine extends AbstractScriptEngine implements C
 
         @Override
         public void flush() throws IOException {
-            writer.flush();
+            if (writer != null) {
+                writer.flush();
+            }
         }
 
         void setWriter(Writer writer) {