+++ graal-nodejs/mx.graal-nodejs/mx_graal_nodejs.py
@@ -308,7 +308,8 @@ def setupNodeEnvironment(args):
         _setEnvVar('TRUFFLE_JAR_PATH', mx.distribution('truffle:TRUFFLE_API').path)
     _setEnvVar('LAUNCHER_COMMON_JAR_PATH', mx.distribution('sdk:LAUNCHER_COMMON').path)
     _setEnvVar('TRUFFLENODE_JAR_PATH', mx.distribution('TRUFFLENODE').path)
-    _setEnvVar('NODE_JVM_CLASSPATH', mx.classpath(['tools:CHROMEINSPECTOR', 'tools:TRUFFLE_PROFILER', 'TRUFFLENODE']))
+    _setEnvVar('NODE_JVM_CLASSPATH', mx.classpath(['TRUFFLENODE']
+            + (['tools:CHROMEINSPECTOR', 'tools:TRUFFLE_PROFILER'] if mx.suite('tools', fatalIfMissing=False) is not None else [])))
     setLibraryPath()
 
     prevPATH = os.environ['PATH']
@@ -494,7 +495,7 @@ def buildSvmImage(args):
     for _lang in ['js', 'nodejs']:
         _svm.fetch_languages(['--language:{}=version={}'.format(_lang, _js_version)])
     _svm.fetch_languages(['--tool:regex'])
-    _svm.native_image_on_jvm(['--language:nodejs', '-H:JNIConfigurationResources=svmnodejs.jniconfig'] + args)
+    _svm.native_image_on_jvm(['-H:+EnforceMaxRuntimeCompileMethods', '--language:nodejs', '-H:JNIConfigurationResources=svmnodejs.jniconfig'] + args)
 
 def _prepare_svm_env():
     setLibraryPath()